push5

+++ library/blockmma.h
@@ -44,9 +44,9 @@ extern "C" {
 #include <stdio.h>
 #include <stdlib.h>
 
-int blockmma(int devfd, float *a, float *b, float *c, int m, int n, int k);
-int blockmma_bonus(int devfd, float *a, float *b, float *c, int m, int n, int k);
-int blockmma_f128(int devfd, float *a, float *b, float *c, int m, int n, int k, int tile);
+int blockmma(int devfd, int *a, int *b, int *c, int m, int n, int k);
+int blockmma_bonus(int devfd, int *a, int *b, int *c, int m, int n, int k);
+int blockmma_f128(int devfd, int *a, int *b, int *c, int m, int n, int k, int tile);
 int blockmma_sync(int devfd);
 int blockmma_f128_accelerator(int devfd);
 //int blockmma_free(int devfd, __u64 offset);
